Added 2014-08-20-bluesteel.html

+++ process-content.sh
@@ -3,19 +3,38 @@
 readonly PROGNAME=$(basename $0)
 readonly PROGDIR=$(readlink -m $(dirname $0))
 readonly ARGS="$@"
-readonly TARGET="data"
+GIVEN_TARGET_DIR="$1"
 
-cd $PROGDIR
-rm -rf $TARGET
-cp -ra content $TARGET
-for file in $(find $TARGET -type f -name "*.jpg" -o -name "*.png" -o -name "*.gif"); do
+function processImage() {
+  file=$1
   filename=$(basename $file)
   extension="${filename##*.}"
   filename="${filename%.*}"
-  dir=`dirname $file`
-  echo "Processing $file ..."
-  convert -resize 300x180^ -gravity Center -crop 300x180+0+0 +repage $file $dir/${filename}.crop.$extension
-  convert -resize "500x400>" $file $dir/${filename}.aside.$extension
-  convert -resize "1200x800>" $file $dir/${filename}.large.$extension
-done
+  targetDir=$(readlink -m $(dirname $file))
+  targetDir=$(echo $targetDir | sed "s#$PROGDIR/content/#$PROGDIR/data/#")
+  echo "Processing $file"
+  convert -resize 300x180^ -gravity Center -crop 300x180+0+0 +repage $file $targetDir/${filename}.crop.$extension
+  convert -resize "500x400>" $file $targetDir/${filename}.aside.$extension
+  convert -resize "1200x800>" $file $targetDir/${filename}.large.$extension
+}
 
+cd $PROGDIR
+
+if [ -d "$GIVEN_TARGET_DIR" ]; then
+  GIVEN_TARGET_DIR=$(readlink -m $GIVEN_TARGET_DIR)
+  if [[ "$GIVEN_TARGET_DIR" == "$PROGDIR/content"* ]]; then
+    for image in $(find $GIVEN_TARGET_DIR -type f -name "*.jpg" -o -name "*.png" -o -name "*.gif"); do
+      processImage $image;
+    done
+  else
+    echo "$GIVEN_TARGET_DIR is not valid content-dir"
+    exit 1
+  fi
+else
+  TARGET="data"
+  rm -rf $TARGET
+  cp -ra content $TARGET
+  for image in $(find $TARGET -type f -name "*.jpg" -o -name "*.png" -o -name "*.gif"); do
+    processImage $image
+  done
+fi
